Added `SpotifyAPI.authorizationManagerDidDeauthorize`. This publisher emits when the `deauthorize()` method of the authorization manager is called; `SpotifyAPI.authorizationManagerDidChange` no longer emits when this method is called.
Added the `didDeauthorize` methods to `SpotifyAuthorizationManager` and the conforming authorization managers.

Fixed bug in which creating an instance of `SpotifyAPI` using `init(from:)` did not properly setup the subscription to `authorizationManager.didChange`.

Added tests for encoding and decoding `SpotifyAPI` and ensuring the subscriptions are properly setup. Also added tests for reassigning the authorization manager property of `SpotifyAPI`.